# lxdweb/docs/swagger.yaml
basePath: /
definitions:
...
  models.CreateNATRequest:
    properties:
      container_hostname:
        type: string
      description:
        type: string
      external_port:
        type: integer
      internal_port:
        type: integer
      node_id:
        type: integer
      protocol:
        enum:
        - tcp
        - udp
        type: string
    required:
    - container_hostname
    - external_port
    - internal_port
    - node_id
    - protocol
    type: object
...
  models.UpdateNATRequest:
    properties:
      description:
        type: string
      external_port:
        type: integer
      internal_port:
        type: integer
      protocol:
        enum:
        - tcp
        - udp
        type: string
      status:
        enum:
        - active
        - inactive
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: xkatld@example.com
    name: xkatld
    url: https://github.com/xkatld/zjmf-lxd-server
  description: 基于 Session 的认证
  license:
    name: MIT
    url: https://github.com/xkatld/zjmf-lxd-server/blob/main/LICENSE
  termsOfService: https://github.com/xkatld/zjmf-lxd-server
  title: LXD Web 管理平台
  version: 1.0.3
paths:
...
  /api/nat:
    get:
      description: 查询所有NAT端口转发规则，支持按节点过滤
      parameters:
      - description: 节点ID
        in: query
        name: node_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回NAT规则列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 查询失败
          schema:
            additionalProperties: true
            type: object
      summary: 获取NAT规则列表
      tags:
      - NAT管理
    post:
      consumes:
      - application/json
      description: 创建新的NAT端口转发规则
      parameters:
      - description: NAT规则参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateNATRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 节点不存在
          schema:
            additionalProperties: true
            type: object
        "409":
          description: 端口已被占用
          schema:
            additionalProperties: true
            type: object
      summary: 创建NAT规则
      tags:
      - NAT管理
  /api/nat-sync/all:
    post:
      description: 启动所有活跃节点的NAT规则同步任务
      produces:
      - application/json
      responses:
        "200":
          description: 同步任务已启动
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未登录
          schema:
            additionalProperties: true
            type: object
      summary: 同步所有节点的NAT规则
      tags:
      - NAT管理
  /api/nat-sync/node/{id}:
    post:
      description: 启动指定节点的NAT规则同步任务
      parameters:
      - description: 节点ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 同步任务已启动
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 节点ID格式错误或节点正在同步
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未登录
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 节点不存在
          schema:
            additionalProperties: true
            type: object
      summary: 同步指定节点的NAT规则
      tags:
      - NAT管理
  /api/nat-sync/status:
    get:
      description: 查询指定节点或所有节点的NAT规则同步状态
      parameters:
      - description: 节点ID
        in: query
        name: node_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回同步状态
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未登录
          schema:
            additionalProperties: true
            type: object
      summary: 获取NAT同步状态
      tags:
      - NAT管理
  /api/nat-sync/tasks:
    get:
      description: 查询最近50条NAT规则同步任务记录
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回任务列表
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未登录
          schema:
            additionalProperties: true
            type: object
      summary: 获取NAT同步任务列表
      tags:
      - NAT管理
  /api/nat/{id}:
    delete:
      description: 删除指定的NAT端口转发规则并清理数据库
      parameters:
      - description: 规则ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties: true
            type: object
        "404":
          description: NAT规则或节点不存在
          schema:
            additionalProperties: true
            type: object
      summary: 删除NAT规则
      tags:
      - NAT管理
    get:
      description: 根据ID获取NAT规则详情
      parameters:
      - description: 规则ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回NAT规则
          schema:
            additionalProperties: true
            type: object
        "404":
          description: NAT规则不存在
          schema:
            additionalProperties: true
            type: object
      summary: 获取单个NAT规则
      tags:
      - NAT管理
    put:
      consumes:
      - application/json
      description: 更新NAT规则的描述或状态
      parameters:
      - description: 规则ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateNATRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: NAT规则不存在
          schema:
            additionalProperties: true
            type: object
      summary: 更新NAT规则
      tags:
      - NAT管理
  /api/nat/cache:
    get:
      description: 直接从本地缓存数据库读取NAT规则信息
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回缓存数据
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未登录
          schema:
            additionalProperties: true
            type: object
      summary: 从缓存获取NAT规则列表
      tags:
      - NAT管理
  /api/nat/check:
    get:
      description: 检查指定端口和协议是否可用于NAT转发
      parameters:
      - description: 节点ID
        in: query
        name: node_id
        required: true
        type: string
      - description: 容器名称
        in: query
        name: hostname
        required: true
        type: string
      - description: 协议类型(tcp/udp)
        in: query
        name: protocol
        required: true
        type: string
      - description: 端口号
        in: query
        name: port
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 检查结果
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 节点不存在
          schema:
            additionalProperties: true
            type: object
      summary: 检查NAT端口是否可用
      tags:
      - NAT管理
  /api/nat/sync:
    post:
      description: 同步指定节点的NAT规则信息
      parameters:
      - description: 节点ID
        in: query
        name: node_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 同步任务已启动
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 缺少参数
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 节点不存在
          schema:
            additionalProperties: true
            type: object
      summary: 同步NAT规则
      tags:
      - NAT管理
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"